chore: rename `reducer.ts` to `store.ts`
